# api/src/main/resources/api-definition.yaml
openapi: 3.0.3
info:
  contact:
    name: Sandro Andreasik
  description: Sprint Planning Poker for Scrum teams
  title: EFI Poker API
  version: 1.0.0
servers:
- description: API base path
  url: /api/v1
...
paths:
...
  /projects/{slug}:
    delete:
      operationId: deleteProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Project deleted
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Delete project and all associated data (admin)
      tags:
      - projects
    get:
      operationId: getProjectBySlug
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
          description: Project info
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Get project info
      tags:
      - projects
    patch:
      operationId: updateProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAdminResponse"
          description: Updated project
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Update project (admin)
      tags:
      - projects
...
  /projects/{slug}/rooms:
    get:
      operationId: listRooms
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Participant-Id
        required: false
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/RoomResponse"
                type: array
          description: List of rooms
      summary: List rooms for project
      tags:
      - rooms
    post:
      operationId: createRoom
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
          description: Room created
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Create a room
      tags:
      - rooms
...
  /projects/{slug}/participants/{participantId}:
    delete:
      operationId: deleteParticipant
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: path
        name: participantId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Participant deleted
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Participant not found
      summary: Delete participant (admin only)
      tags:
      - participants
    get:
      operationId: getParticipant
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: path
        name: participantId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantResponse"
          description: Participant info
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Participant not found
      summary: Get participant by ID (magic link validation)
      tags:
      - participants
    patch:
      operationId: updateParticipant
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: path
        name: participantId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Participant-Id
        required: false
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParticipantRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantResponse"
          description: Updated participant
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not allowed to update this participant
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Participant not found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Nickname already taken in this project
      summary: Update participant nickname (self-service)
      tags:
      - participants
  /rooms/by-slug/{roomSlug}:
    get:
      operationId: getRoomBySlug
      parameters:
      - in: path
        name: roomSlug
        required: true
        schema:
          pattern: "^[A-Z0-9]{3}-[A-Z0-9]{3}$"
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSlugResponse"
          description: Room lookup result
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Room not found
      summary: Look up room by slug code
      tags:
      - rooms
  /rooms/{roomId}:
    get:
      operationId: getRoom
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Participant-Id
        required: false
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailResponse"
          description: Room details
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Get room with tasks and own estimates
      tags:
      - rooms
    patch:
      operationId: updateRoom
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
          description: Updated room
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Update room
      tags:
      - rooms
  /rooms/{roomId}/admin:
    get:
      operationId: getRoomAdmin
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomAdminResponse"
          description: Room admin view
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Get room with ALL estimates (admin)
      tags:
      - rooms
  /rooms/{roomId}/reveal:
    post:
      operationId: revealEstimates
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailResponse"
          description: Estimates revealed
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Force reveal all estimates
      tags:
      - rooms
  /rooms/{roomId}/reopen:
    post:
      operationId: reopenRoom
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetailResponse"
          description: Room reopened
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
      summary: Reopen a revealed or closed room
      tags:
      - rooms
  /rooms/{roomId}/progress:
    get:
      operationId: getRoomProgress
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomProgressResponse"
          description: Progress info
      summary: Get voting progress (X/Y per task)
      tags:
      - rooms
  /rooms/{roomId}/results:
    get:
      operationId: getRoomResults
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResultsResponse"
          description: Full results
        "403":
          description: Not yet revealed
      summary: Get full results (only if REVEALED/CLOSED)
      tags:
      - rooms
  /rooms/{roomId}/results/export:
    get:
      operationId: exportResults
      parameters:
      - in: path
        name: roomId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            text/csv:
              schema:
                type: string
          description: CSV export
      summary: Export results as CSV
      tags:
      - rooms
...
components:
...
  schemas:
...
    ProjectAdminResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        adminCode: adminCode
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        slug:
          type: string
        adminCode:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
    ProjectResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        slug:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
    ProblemDetail:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      type: object
    UpdateProjectRequest:
      example:
        name: name
      properties:
        name:
          maxLength: 255
          minLength: 1
          type: string
      type: object
    RoomResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        roundNumber: 0
        description: description
        topic: topic
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        deadline: 2000-01-23T04:56:07.000+00:00
        projectId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
        roomType: ASYNC
        status: OPEN
      properties:
        id:
          format: uuid
          type: string
        slug:
          description: Human-readable room code (e.g. A3X-K7B)
          type: string
        projectId:
          format: uuid
          type: string
        title:
          type: string
        description:
          type: string
        roomType:
          $ref: "#/components/schemas/RoomType"
        deadline:
          format: date-time
          type: string
        status:
          $ref: "#/components/schemas/RoomStatus"
        topic:
          type: string
        roundNumber:
          type: integer
        createdAt:
          format: date-time
          type: string
      type: object
...
    RoomType:
      enum:
      - ASYNC
      - LIVE
      type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http